added pcap parsing demonstration using dpkt
Note that using the same code on 'test.pcap' fails, likely because the
packets in test.pcap are using some other kind of protocol.
(specifically, running it causes eth.data to return a string type,
causing ip.data to raise an error since .data is not a method for
strings).